refactor: Location 파일들 패키지로 분리
