+++ app/src/main/java/com/bestapp/rice/Location/LocationViewModel.kt
@@ -1,4 +1,4 @@
-package com.bestapp.rice.ui.meetupmap
+package com.bestapp.rice.Location
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
