+++ app/src/main/java/com/bestapp/rice/ui/meetupmap/MeetUpMapFragment.kt
@@ -12,6 +12,7 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
+import com.bestapp.rice.Location.LocationViewModel
 import com.bestapp.rice.R
 import com.bestapp.rice.databinding.FragmentMeetUpMapBinding
 import com.kakao.vectormap.KakaoMap
